move objects to /fixtures
