+++ cypress/integration/teams.spec.ts
@@ -1,7 +1,7 @@
-import { createUser, userMigWithoutPreferences } from './user.spec';
-import { Team, UserRef } from '../../src/app/types/team';
-import { List } from '../../src/app/types/list';
-import { organizationListNxtVshn } from './organizations.spec';
+import { Team } from '../../src/app/types/team';
+import { createUser, userMigWithoutPreferences } from '../fixtures/user';
+import { createTeamList, team1, teamListNxt, teamListVshn } from '../fixtures/team';
+import { organizationListNxtVshn } from '../fixtures/organization';
 
 describe('Test teams list', () => {
   beforeEach(() => {
@@ -21,7 +21,7 @@ describe('Test teams list', () => {
       body: organizationListNxtVshn,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams', {
-      body: teamsNxt,
+      body: teamListNxt,
     });
     cy.get('#teams-title').should('contain.text', 'Teams');
     cy.get(':nth-child(2) > .flex-row > .text-3xl').should('contain.text', 'team1');
@@ -40,7 +40,7 @@ describe('Test teams list', () => {
       body: organizationListNxtVshn,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/vshn/teams', {
-      body: teamsVshn,
+      body: teamListVshn,
     });
     cy.get('#teams-title').should('contain.text', 'Teams');
     cy.get(':nth-child(2) > .flex-row > .text-3xl').should('contain.text', 'tarazed');
@@ -58,10 +58,10 @@ describe('Test teams list', () => {
       body: organizationListNxtVshn,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/vshn/teams', {
-      body: teamsVshn,
+      body: teamListVshn,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams', {
-      body: teamsNxt,
+      body: teamListNxt,
     });
     cy.get('#teams-title').should('contain.text', 'Teams');
     cy.get(':nth-child(2) > .flex-row > .text-3xl').should('contain.text', 'tarazed');
@@ -127,7 +127,7 @@ describe('Test team edit', () => {
       body: organizationListNxtVshn,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams', {
-      body: teamsNxt,
+      body: teamListNxt,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams/team1', {
       body: team1,
@@ -178,7 +178,7 @@ describe('Test team edit', () => {
       body: organizationListNxtVshn,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams', {
-      body: teamsNxt,
+      body: teamListNxt,
     });
     cy.visit('/teams');
     cy.get('#teams-title').should('contain.text', 'Teams');
@@ -208,7 +208,7 @@ describe('Test teams add', () => {
       body: organizationListNxtVshn,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams', {
-      body: teamsNxt,
+      body: teamListNxt,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams/team1', {
       body: team1,
@@ -259,7 +259,7 @@ describe('Test teams add', () => {
       body: organizationListNxtVshn,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams', {
-      body: teamsNxt,
+      body: teamListNxt,
     });
     cy.visit('/teams');
     cy.get('#teams-title').should('contain.text', 'Teams');
@@ -285,7 +285,7 @@ describe('Test teams delete', () => {
       body: organizationListNxtVshn,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams', {
-      body: teamsNxt,
+      body: teamListNxt,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams/team1', {
       body: team1,
@@ -318,7 +318,7 @@ describe('Test teams delete', () => {
       body: organizationListNxtVshn,
     });
     cy.intercept('GET', 'appuio-api/apis/appuio.io/v1/namespaces/nxt/teams', {
-      body: teamsNxt,
+      body: teamListNxt,
     });
     cy.visit('/teams');
     cy.get('#teams-title').should('contain.text', 'Teams');
@@ -327,76 +327,3 @@ describe('Test teams delete', () => {
     cy.get(':nth-child(2) > .flex-row > :nth-child(2) > [title="Delete team"]').should('not.exist');
   });
 });
-
-export interface TeamConfig {
-  name: string;
-  namespace: string;
-  displayName: string;
-  userRefs: UserRef[];
-}
-
-export interface TeamListConfig {
-  items: Team[];
-}
-
-export const team1 = createTeam({
-  name: 'team1',
-  namespace: 'nxt',
-  displayName: 'My Super Team 1',
-  userRefs: [{ name: 'mig' }, { name: 'miw' }],
-});
-
-export const teamsNxt = createTeamList({
-  items: [
-    createTeam({
-      name: 'team1',
-      namespace: 'nxt',
-      displayName: 'My Super Team 1',
-      userRefs: [{ name: 'mig' }, { name: 'miw' }],
-    }),
-    createTeam({
-      name: 'team2',
-      namespace: 'nxt',
-      displayName: 'My Super Team 2',
-      userRefs: [{ name: 'cma' }],
-    }),
-  ],
-});
-
-export const teamsVshn = createTeamList({
-  items: [
-    createTeam({
-      name: 'tarazed',
-      namespace: 'vshn',
-      displayName: 'Tarazed',
-      userRefs: [{ name: 'corvus' }],
-    }),
-  ],
-});
-
-export function createTeam(teamConfig: TeamConfig): Team {
-  return {
-    apiVersion: 'appuio.io/v1',
-    kind: 'Team',
-    metadata: {
-      name: teamConfig.name,
-      namespace: teamConfig.namespace,
-    },
-    spec: {
-      displayName: teamConfig.displayName,
-      userRefs: teamConfig.userRefs,
-    },
-  };
-}
-
-export function createTeamList(teamListConfig: TeamListConfig): List<Team> {
-  return {
-    apiVersion: 'v1',
-    kind: 'List',
-    items: teamListConfig.items,
-    metadata: {
-      resourceVersion: '',
-      selfLink: '',
-    },
-  };
-}
