+++ cypress/integration/user.spec.ts
@@ -1,5 +1,5 @@
-import { User, UserSpec } from 'src/app/types/user';
-import { organizationListNxtVshn } from './organizations.spec';
+import { createUser, userMigWithoutPreferences } from '../fixtures/user';
+import { organizationListNxtVshn } from '../fixtures/organization';
 
 describe('Test user', () => {
   beforeEach(() => {
@@ -76,30 +76,3 @@ describe('Test user', () => {
     cy.get('.p-dropdown-label').should('contain.text', 'None');
   });
 });
-
-export interface UserConfig {
-  username: string;
-  defaultOrganizationRef?: string;
-}
-
-export const userMigWithoutPreferences = createUser({ username: 'mig' });
-
-export function createUser(userConfig: UserConfig): User {
-  let spec: UserSpec = {};
-  if (userConfig.defaultOrganizationRef) {
-    spec = {
-      preferences: {
-        defaultOrganizationRef: userConfig.defaultOrganizationRef,
-      },
-    };
-  }
-  return {
-    apiVersion: 'appuio.io/v1',
-    kind: 'User',
-    metadata: {
-      name: userConfig.username,
-      resourceVersion: '',
-    },
-    spec,
-  };
-}
